Move /ui to /shared/ui folder in WebApp
