+++ application/account-management/WebApp/shared/ui/LocaleSwitcher.tsx
