+++ application/account-management/WebApp/shared/ui/images/HeroImage.tsx
