+++ application/account-management/WebApp/shared/ui/oneTimePassword/DigitPattern.tsx
